requirements.txt
Django
djangorestframework
markdown
django-filter
django-angular
Pillow
django-rest-registration
